+++ mex/extractors/open_data/transform.py
@@ -36,6 +36,9 @@ from mex.extractors.wikidata.helpers import (
     get_wikidata_extracted_organization_id_by_name,
 )
 
+# TODO @MX-2075: remove
+FALLBACK_UNIT = "mf4"
+
 
 def transform_open_data_person_affiliations_to_organizations(
     open_data_creators_contributors: list[OpenDataCreatorsOrContributors],
@@ -269,9 +272,11 @@ def transform_open_data_parent_resource_to_mex_resource(  # noqa: PLR0913
         for distribution in open_data_distribution
     }
     has_personal_data = resource_mapping.hasPersonalData[0].mappingRules[0].setValues
-    resource_type_general = (
-        resource_mapping.resourceTypeGeneral[0].mappingRules[0].setValues
-    )
+    resource_type_general_lookup = {
+        rule.forValues[0]: rule.setValues
+        for rule in resource_mapping.resourceTypeGeneral[0].mappingRules
+        if rule.forValues
+    }
     theme = resource_mapping.theme[0].mappingRules[0].setValues
     language_by_keyword = {
         rule.forValues[0]: rule.setValues[0]
@@ -279,7 +284,7 @@ def transform_open_data_parent_resource_to_mex_resource(  # noqa: PLR0913
         if rule.forValues and rule.setValues
     }
     for resource in open_data_parent_resource:
-        contributing_unit = list(
+        unit_in_charge = list(
             {
                 unit_id: None
                 for person in resource.metadata.contributors
@@ -292,6 +297,8 @@ def transform_open_data_parent_resource_to_mex_resource(  # noqa: PLR0913
                 for unit_id in unit_list
             }
         )
+        if not unit_in_charge:
+            unit_in_charge = [unit_stable_target_ids_by_synonym[FALLBACK_UNIT]]
         contributor = [
             c
             for person in resource.metadata.contributors
@@ -331,12 +338,15 @@ def transform_open_data_parent_resource_to_mex_resource(  # noqa: PLR0913
             in resource_mapping.license[0].mappingRules[0].forValues
             else None
         )
-        unit_in_charge = (
+        contributing_unit = (
             unit_stable_target_ids_by_synonym.get(
-                resource_mapping.unitInCharge[0].mappingRules[0].forValues[0]
+                resource_mapping.contributingUnit[0].mappingRules[0].forValues[0]
             )
-            if resource_mapping.unitInCharge[0].mappingRules[0].forValues
+            if resource_mapping.contributingUnit[0].mappingRules[0].forValues
             else None
+        )
+        resource_type_general = resource_type_general_lookup.get(
+            resource.metadata.resource_type.type, []
         )
         extracted_resource.append(
             ExtractedResource(