+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - publish all consenting persons in publisher
 - bump mex-common dependency to 1.16.0
 - use primary source id convenience function
+- update open data resource according to mapping changes
 
 ### Deprecated